requirements.txt
# Available in stretch-backports
Django>=1.11.7,<2.0

# Available in Debian
psycopg2
django-countries
python-memcached
PyYAML

# From the cheeseshop
wafer==0.7.2
django-crispy-forms>=1.7.0
django-paypal==0.4.1
Django==1.9.8
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
Unipath==1.0
python-decouple==3
Pillow==3.3.0
Markdown==2.6.6
bleach==1.4.3
psycopg2==2.6.1
whitenoise==3.2
